[r2r] update metrics related dep && refactoring (#1312)
* update metrics related dep

* wasm don't mess up!

* wip

* wip

* impl collect_json done

* added collect_json test

* created new mm2_metrics crate and other wip impls

* update wasm

* rust fmt check

* wip

* MetricsArc && MetricsWeak

* wip

* wip

* removed unused dep

* refactoring

* removed unused deps

* collect gauge from bits to u64

* gauge now export u64

* updated histogram and gauge

* iteration

* reiteration

* reiteration

* Iterations

* histogram, label macro and other improvements

* fixed conflicts

* fix udeps, impl wasm for metrics and other iterations

* code improvements

* improved TryRecorder for MetricsWeak
Added unit test for test_prometheus_format

* renamed native to mm_metrics
unit tests working correctly
code improvements

* fix cargo fmt

* changed incompatible quanta dep --v

* PR review fixes - 1

* PR Review fixes -2

* Fix WASM

* PR review fixes - wip

* PR review fixes

* Changes to error handling and unit tests

* Fixed test_collect_json test deltas

* PR review fixes - 3

* edited lp_commands_legacy

* PR review fix - last

* Minor changes - r2r

* fix cargo clippy

* final pr review changes

* fixed build error

+++ mm2src/mm2_main/src/lp_ordermatch.rs
@@ -42,6 +42,7 @@ use keys::{AddressFormat, KeyPair};
 use mm2_core::mm_ctx::{from_ctx, MmArc, MmWeak};
 use mm2_err_handle::prelude::*;
 use mm2_libp2p::{decode_signed, encode_and_sign, encode_message, pub_sub_topic, TopicPrefix, TOPIC_SEPARATOR};
+use mm2_metrics::{mm_gauge, mm_label};
 use mm2_number::{construct_detailed, BigDecimal, BigRational, Fraction, MmNumber, MmNumberMultiRepr};
 #[cfg(test)] use mocktopus::macros::*;
 use num_traits::identities::Zero;
@@ -2380,13 +2381,13 @@ fn collect_orderbook_metrics(ctx: &MmArc, orderbook: &Orderbook) {
     }
 
     let memory_db_size = malloc_size(&orderbook.memory_db);
-    mm_gauge!(ctx.metrics, "orderbook.len", orderbook.order_set.len() as i64);
-    mm_gauge!(ctx.metrics, "orderbook.memory_db", memory_db_size as i64);
+    mm_gauge!(ctx.metrics, "orderbook.len", orderbook.order_set.len() as f64);
+    mm_gauge!(ctx.metrics, "orderbook.memory_db", memory_db_size as f64);
 
     // TODO remove metrics below after testing
     for (pubkey, pubkey_state) in orderbook.pubkeys_state.iter() {
-        mm_gauge!(ctx.metrics, "orders_uuids", pubkey_state.orders_uuids.len() as i64, "pubkey" => pubkey.clone());
-        mm_gauge!(ctx.metrics, "history.commited_changes", history_committed_changes(&pubkey_state.order_pairs_trie_state_history), "pubkey" => pubkey.clone());
+        mm_gauge!(ctx.metrics, "orders_uuids", pubkey_state.orders_uuids.len() as f64, "pubkey" => pubkey.clone());
+        mm_gauge!(ctx.metrics, "history.commited_changes", history_committed_changes(&pubkey_state.order_pairs_trie_state_history) as f64, "pubkey" => pubkey.clone());
     }
 }
